+++ python/SEGMENTACION DE CLIENTES/JSONprueba.py
@@ -55,17 +55,17 @@ def iniciarPrograma(archivo):
         print(archivo)
         archivo = eventos_classic
         cargado = Json(archivo)
-        return Cliente_clasico(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
+        return Cliente_clasico(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo), cargado
     elif archivo == "GOLD":
         print(archivo)
         archivo = eventos_gold
         cargado = Json(archivo)
-        return Cliente_gold(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
+        return Cliente_gold(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo), cargado
     elif archivo == "BLACK":
         print(archivo)
         archivo = eventos_black
         cargado = Json(archivo)
-        return Cliente_black(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
+        return Cliente_black(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo), cargado
     else:
         print("archivo no reconocido")
