+++ python/SEGMENTACION DE CLIENTES/indexpy.html
@@ -1,22 +1,29 @@
 <html>
     <head>
     <title>INFORME CLIENTE</title>
-    <h1>CLIENTE NR: 1541425 </h1>    <h1>NOMBRE:Nicolas</h1>   <h1>APELLIDO:Gaston</h1>      <h1> DNI:29494777 </h1>
+    <h1>CLIENTE NR: Garcia </h1>    <h1>NOMBRE:Mariano</h1>   <h1>APELLIDO:Garcia</h1>      <h1> DNI:1546859 </h1>
     <h1>DIRECCION:</h1><p>PAIS: Argentina</p>
     <p>PROVINCIA: Buenos Aires</p>
     <p>CIUDAD: Capital Federal</p>
-    <p>CALLE: Uriburu</p>
-    <p>NUMERO: 8700</p>
+    <p>CALLE: Bolivia</p>
+    <p>NUMERO: 7900</p>
     </head>
     <body>
     <p>Se realizaron las siguientes operaciones:</p>
 
-    <p>0 | 23/06/2022 16:00:55 | ALTA_CHEQUERA | CUENTA NUMERO 190 | RECHAZADA | motivo: limite de chequeras alcanzado, limite: 0 chequeras actuales: 0</p>
-<p>1 | 22/06/2022 16:00:55 | ALTA_TARJETA_CREDITO | CUENTA NUMERO 190 | RECHAZADA | motivo: limte de tarjetas de credito alcanzado, limite: 0 tarjetas actuales: 0</p>
-<p>2 | 24/06/2022 16:00:55 | COMPRA_DOLAR | CUENTA NUMERO 190 | RECHAZADA | motivo: cliente no habilitado para comprar dolares, debe ser BLACK o GOLD</p><p>3 | 20/06/2022 16:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | ACEPTADA | Cupo diario restante: 9000 | monto solicitado: 1000 </p>
-<p>4 | 20/06/2022 17:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: El monto solicitado excede el cupo diario: Cupo diario restante: 9000 | monto solicitado: 11000 </p>
-<p>5 | 20/06/2022 18:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: El monto solicitado excede el cupo diario: Cupo diario restante: 3000 | monto solicitado: 9000 </p>
-<p>6 | 21/06/2022 10:11:11 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: El monto solicitado excede el cupo diario: Cupo diario restante: 3000 | monto solicitado: 9000 </p>
-<p>7 | 25/06/2022 16:00:55 | TRANSFERENCIA_ENVIADA | CUENTA NUMERO 190 | RECHAZADA | motivo: Saldo insuficiente, en las cuentas CLASSIC el saldo no puede quedar negativo | saldo restante: 0</p>
-<p>8 | 24/06/2022 17:00:00 | TRANSFERENCIA_RECIBIDA | CUENTA NUMERO 190 | ACEPTADA | monto de transferencia: 9000</p>
-<p>9 | 25/06/2022 16:00:55 | TRANSFERENCIA_RECIBIDA | CUENTA NUMERO 190 | RECHAZADA | motivo: Limite de transferecia recibida excedido, debe pedir autorizacion al banco para realizar transferencias tan grandes | monto de transferencia: 200000</p>
+    <p>0 | 06/06/2022 10:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | ACEPTADA | Cupo diario restante: 9000 | monto solicitado: 1000 </p>
+<p>1 | 06/06/2022 10:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | ACEPTADA | Cupo diario restante: 9000 | monto solicitado: 1000 </p>
+<p>2 | 06/06/2022 10:10:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: Cupo diario de extraccion superado, no puede superar los $10000 por dia: Cupo diario restante: 10000 | monto solicitado: 11000 </p>
+<p>3 | 06/06/2022 10:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | ACEPTADA | Cupo diario restante: 9000 | monto solicitado: 1000 </p>
+<p>4 | 06/06/2022 10:10:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: Cupo diario de extraccion superado, no puede superar los $10000 por dia: Cupo diario restante: 10000 | monto solicitado: 11000 </p>
+<p>5 | 06/06/2022 10:22:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: Cupo diario de extraccion superado, no puede superar los $3000 por dia: Cupo diario restante: 3000 | monto solicitado: 9000 </p>
+<p>6 | 06/06/2022 10:00:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | ACEPTADA | Cupo diario restante: 9000 | monto solicitado: 1000 </p>
+<p>7 | 06/06/2022 10:10:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: Cupo diario de extraccion superado, no puede superar los $10000 por dia: Cupo diario restante: 10000 | monto solicitado: 11000 </p>
+<p>8 | 06/06/2022 10:22:55 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: Cupo diario de extraccion superado, no puede superar los $3000 por dia: Cupo diario restante: 3000 | monto solicitado: 9000 </p>
+<p>9 | 06/06/2022 10:33:33 | RETIRO_EFECTIVO_CAJERO_AUTOMATICO | CUENTA NUMERO 190 | RECHAZADA | motivo: Cupo diario de extraccion superado, no puede superar los $3000 por dia: Cupo diario restante: 3000 | monto solicitado: 9000 </p>
+<p>10 | 06/06/2022 12:00:00 | ALTA_TARJETA_CREDITO | CUENTA NUMERO 190 | RECHAZADA | motivo: ha superado el limte de tarjetas de credito, el limite de esta cuenta es 5 tarjetas actuales: 5</p>
+<p>11 | 06/06/2022 12:30:55 | ALTA_CHEQUERA | CUENTA NUMERO 190 | RECHAZADA | motivo: ha superado el limte de chequeras de su cuenta, el limite es 2 chequeras actuales: 2</p>
+<p>12 | 06/06/2022 12:45:33 | COMPRA_DOLAR | CUENTA NUMERO 190 | ACEPTADA | </p><p>13 | 06/06/2022 13:00:55 | TRANSFERENCIA_ENVIADA | CUENTA NUMERO 190 | RECHAZADA | motivo: Saldo en cuenta insuficiente | saldo restante: 900000</p>
+<p>14 | 06/06/2022 13:11:15 | TRANSFERENCIA_RECIBIDA | CUENTA NUMERO 190 | ACEPTADA | monto de transferencia: 9000</p>
+<p>15 | 06/06/2022 13:11:15 | TRANSFERENCIA_RECIBIDA | CUENTA NUMERO 190 | ACEPTADA | monto de transferencia: 9000</p>
+<p>16 | 06/06/2022 16:00:55 | TRANSFERENCIA_RECIBIDA | CUENTA NUMERO 190 | ACEPTADA | monto de transferencia: 200000</p>
