+++ python/SEGMENTACION DE CLIENTES/print_HTML.py
@@ -1,27 +1,28 @@
-from razones import RazonAltaChequera,RazonAltaTarjetaCredito,RazonCompraDolar,RazonRetiroEfectivo,RazonTransferenciaEnviada,RazonTransferenciaRecibida
+from razones import razones
 #variable que va a almacenar todas las transacciones
 transeferencias = None
 
 #funcion que junta todas las transacciones con sus razones de rechazo ya especificadas
-def obtenerRazones(cliente):
-    #invocamos las clases
-    RazonAltaChequera().resolver(cliente)
-    RazonTransferenciaRecibida().resolver(cliente)
-    RazonTransferenciaEnviada().resolver(cliente)
-    RazonRetiroEfectivo().resolver(cliente)
-    RazonCompraDolar().resolver(cliente)
-    RazonAltaTarjetaCredito().resolver(cliente)
-    #juntamos los valores de las listas
-    transeferencias = (RazonAltaChequera().rechazados + RazonAltaChequera().aprobados)
-    transeferencias +=(RazonAltaTarjetaCredito().rechazados + RazonAltaTarjetaCredito().aprobados)
-    transeferencias += (RazonCompraDolar().aprobados + RazonCompraDolar().rechazados)
-    transeferencias += (RazonRetiroEfectivo().aprobados + RazonRetiroEfectivo().rechazados)
-    transeferencias +=(RazonTransferenciaEnviada().rechazados + RazonTransferenciaEnviada().aprobados)
-    transeferencias +=(RazonTransferenciaRecibida().aprobados+ RazonTransferenciaRecibida().rechazados)
-    return transeferencias
+# def obtenerRazones(cliente):
+#     #invocamos las clases
+#     RazonAltaChequera().resolver(cliente)
+#     RazonTransferenciaRecibida().resolver(cliente)
+#     RazonTransferenciaEnviada().resolver(cliente)
+#     RazonRetiroEfectivo().resolver(cliente)
+#     RazonCompraDolar().resolver(cliente)
+#     RazonAltaTarjetaCredito().resolver(cliente)
+#     #juntamos los valores de las listas
+#     transeferencias = (RazonAltaChequera().rechazados + RazonAltaChequera().aprobados)
+#     transeferencias +=(RazonAltaTarjetaCredito().rechazados + RazonAltaTarjetaCredito().aprobados)
+#     transeferencias += (RazonCompraDolar().aprobados + RazonCompraDolar().rechazados)
+#     transeferencias += (RazonRetiroEfectivo().aprobados + RazonRetiroEfectivo().rechazados)
+#     transeferencias +=(RazonTransferenciaEnviada().rechazados + RazonTransferenciaEnviada().aprobados)
+#     transeferencias +=(RazonTransferenciaRecibida().aprobados+ RazonTransferenciaRecibida().rechazados)
+#     return transeferencias
 
 
 def conversion(lista):
+    
     for x in enumerate(lista):
         with open("python\SEGMENTACION DE CLIENTES\indexpy.html", "a+") as e:
             if x[1]['estado'] == 'RECHAZADA':
@@ -53,7 +54,7 @@ def conversion(lista):
                     e.write (f"""<p>{x[0]} | {x[1]['fecha']} | {x[1]['tipo']} | CUENTA NUMERO {x[1]['cuentaNumero']} | {x[1]['estado']} | monto de transferencia: {x[1]['monto']}</p>\n""")
                     
 
-def plantillaHTML(cliente):
+def plantillaHTML(cliente,transacciones):
     archivoHTML = open("python\SEGMENTACION DE CLIENTES\indexpy.html", "w")
     htmlBase = f"""<html>
     <head>
@@ -71,5 +72,5 @@ def plantillaHTML(cliente):
     """
     archivoHTML.write(htmlBase)
     archivoHTML.close()
-    conversion(obtenerRazones(cliente))
+    conversion(razones(cliente = cliente,archivo = transacciones))
